调整 KMongo 模块逻辑

+++ module-kotlin/src/main/kotlin/com/zmkn/module/kmongo/extension/CoroutinePublisher.kt
@@ -395,7 +395,7 @@ fun <T : Any> CoroutineCollection<T>.projection(
                         if (it.path() == "_id") {
                             existId = true
                         }
-                        newProjectionMap[it.path()] = value as Int
+                        newProjectionMap[it.path()] = value
                         Pair(
                             it.path(),
                             it.returnType.classifier as KClass<*>
